src/requirements.txt
pyyaml
dateparser
pytest
lxml
pillow
beautifulsoup4
black
Unidecode
mutagen
internetarchive
eyed3
google-api-python-client
openapi
wandb

